dl/translation: move record_mux/translation_task to translation/
No logical changes.

Moves translation related stuff like record multipler and
translation_task into src/v/datalake/translation/[tests] for better
dependency hierarchy. Needed for a future commit that adds certain
exceptions specific to translation and it is a bit odd to have stuff in
src/v/datalake/ depend on these new types in translation. Also these
utilities are only used with translation code which makes this move
logical.
