+++ src/v/datalake/translation/BUILD
@@ -85,6 +85,71 @@ redpanda_cc_library(
 )
 
 redpanda_cc_library(
+    name = "record_multiplexer",
+    srcs = [
+        "record_multiplexer.cc",
+    ],
+    hdrs = [
+        "record_multiplexer.h",
+    ],
+    implementation_deps = [
+        "//src/v/base",
+        "//src/v/datalake:catalog_schema_manager",
+        "//src/v/iceberg/conversion:conversion_outcome",
+        "//src/v/datalake:logger",
+        "//src/v/datalake:record_schema_resolver",
+        "//src/v/datalake:record_translator",
+        "//src/v/datalake:table_creator",
+        "//src/v/datalake:table_id_provider",
+        "//src/v/storage:parser_utils",
+        "//src/v/utils:lazy_abort_source",
+        "//src/v/utils:prefix_logger",
+    ],
+    include_prefix = "datalake/translation",
+    visibility = [":__subpackages__"],
+    deps = [
+        "//src/v/container:chunked_hash_map",
+        "//src/v/datalake:location",
+        "//src/v/datalake:partitioning_writer",
+        "//src/v/datalake:schema_identifier",
+        "//src/v/datalake:types",
+        "//src/v/datalake:writer",
+        "//src/v/datalake/translation:translation_probe",
+        "//src/v/model",
+        "@seastar",
+    ],
+)
+
+redpanda_cc_library(
+    name = "translation_task",
+    srcs = [
+        "translation_task.cc",
+    ],
+    hdrs = [
+        "translation_task.h",
+    ],
+    include_prefix = "datalake/translation",
+    visibility = ["//visibility:public"],
+    deps = [
+        ":record_multiplexer",
+        "//src/v/base",
+        "//src/v/datalake:cloud_data_io",
+        "//src/v/datalake:location",
+        "//src/v/datalake:logger",
+        "//src/v/datalake:types",
+        "//src/v/datalake:writer",
+        "//src/v/datalake/coordinator:translated_offset_range",
+        "//src/v/datalake/translation:translation_probe",
+        "//src/v/iceberg:values_bytes",
+        "//src/v/model",
+        "//src/v/utils:lazy_abort_source",
+        "//src/v/utils:retry_chain_node",
+        "@fmt",
+        "@seastar",
+    ],
+)
+
+redpanda_cc_library(
     name = "translator_deps",
     srcs = [
         "deps.cc",
@@ -98,6 +163,7 @@ redpanda_cc_library(
         ":scheduler",
         ":stm",
         ":translation_probe",
+        ":translation_task",
         ":utils",
         "//src/v/base",
         "//src/v/cluster",
@@ -108,7 +174,6 @@ redpanda_cc_library(
         "//src/v/datalake:record_translator",
         "//src/v/datalake:serde_parquet_writer",
         "//src/v/datalake:table_creator",
-        "//src/v/datalake:translation_task",
         "//src/v/datalake:types",
         "//src/v/datalake:writer",
         "//src/v/datalake/coordinator:frontend",
