+++ src/v/datalake/translation/deps.cc
@@ -20,7 +20,7 @@
 #include "datalake/serde_parquet_writer.h"
 #include "datalake/translation/state_machine.h"
 #include "datalake/translation/utils.h"
-#include "datalake/translation_task.h"
+#include "datalake/translation/translation_task.h"
 #include "kafka/data/partition_proxy.h"
 #include "kafka/utils/txn_reader.h"
 #include "utils/human.h"
