+++ src/v/datalake/translation/tests/record_multiplexer_bench.cc
@@ -12,13 +12,13 @@
 #include "container/fragmented_vector.h"
 #include "datalake/catalog_schema_manager.h"
 #include "datalake/location.h"
-#include "datalake/record_multiplexer.h"
 #include "datalake/record_schema_resolver.h"
 #include "datalake/record_translator.h"
 #include "datalake/tests/catalog_and_registry_fixture.h"
 #include "datalake/tests/record_generator.h"
 #include "datalake/tests/test_data_writer.h"
 #include "datalake/tests/test_utils.h"
+#include "datalake/translation/record_multiplexer.h"
 #include "model/compression.h"
 #include "model/record.h"
 #include "model/record_batch_reader.h"
