+++ src/v/datalake/translation/translation_task.h
@@ -15,7 +15,7 @@
 #include "datalake/data_writer_interface.h"
 #include "datalake/fwd.h"
 #include "datalake/location.h"
-#include "datalake/record_multiplexer.h"
+#include "datalake/translation/record_multiplexer.h"
 #include "datalake/translation/translation_probe.h"
 #include "model/metadata.h"
 #include "model/record_batch_reader.h"
